Add prop backdropMode to Minicart
