+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `backdropMode` on `minicart.v2`.
+
+### Security
+-  Bump dependency versions.
 
 ## [2.45.0] - 2020-03-25
 ### Added
