+++ react/typings/vtex.store-drawer.d.ts
@@ -1,0 +1,4 @@
+declare module 'vtex.store-drawer' {
+  const Drawer: React.FC<any>
+  type BackdropMode = 'visible' | 'none'
+}
