--- react/typings/vtex.store-drawer.ts
@@ -1,1 +1,0 @@
-declare module 'vtex.store-drawer'
